Correctly route db operation on the right db.
We need to correctly allow or not migration depending on witch db we are
currently working on.

The allow_migrate does not have access to the model but to the
app_label and model_name. So we change the way we decide how we route.

The information is not in the model anymore but directly in a dict.
We get the right info by looking in the dict using the key
(app_label, model_name).