Better handle the library to stock migration
One of the things happening during this migration is that the serial
attribute of the old BookSpecimen model is becoming the barcode
attribute of the Specimen model.

Before this migration, people had to enter their books and specimens
both in the stock (StockItem/Specimen) and in the library
(Book/BookSpecimen).

Because they had to do it in both places, it is very possible that they
had a Specimen with a certain barcode, and a BookSpecimen with the same
serial. (see the first paragraph for how this field maps)

This crashes the migration, because Specimen.barcode has a unicity
constraint.

This commit ensures that the migration still works, ffectively merging
the existing StockItem and Specimen with the new Book and BookSpecimen.

Fixes #626