Workaround the bad Koombook battery driver
This is pretty ugly, but the Koombook battery driver doesn't respect the
Linux convention for reporting the power status to sysfs.

As a result, batinfo doesn't work with it.

Since this is quite specific, I don't think it makes sense to push that
upstream to batinfo, so I opted for a workaround in our code.

Fixes #446