Pretend this migration is reversible
In order to test the migration, we need to be able to revert them first,
since Django "helpfully" runs all migrations before letting us test
anything.

But since we don't want to actually support data rollbacks in Ideascube,
using a noop achieves the purpose needed for the tests, without going
through the trouble of implementing a proper revert.