users: Add an 'extra' field
This is a free-form text field, which can be used for anything at all,
depending on the needs of each deployment.

However, it should only be used for things that are too specific to put
in a dedicated field shared by all deployments.

Like every other user fields, this new one can be shown in forms and
exported in CSVs by adding it to the `USER_FORM_FIELD` setting.

Additionally, its user-facing name can be controlled with the
`USER_EXTRA_FIELD_LABEL` setting.