Fix the migration
This migration was introduced in c06ca9b76a37e1c604a8a5cba4973ebb07a41dea.

However, we modified the model code as part of the code review process,
but forgot to adapt this migration in consequence.

Since we haven't made a release yet, we can afford to modify this
migration in place, rather than making a new one.