Allow adding new user import file formats
This commit introduces a new USER_IMPORT_FORMATS configuration option,
which deployments can specify to a list of formats to use for the
deployment.

Of course, those formats must be supported by the importer, and for now
only the 'ideascube' format is supported. (i.e the format generated by
exporting the users in Ideascube)

If the option is not specified, then only the 'ideascube' format is
available to import from, and therefore the field is hidden.

Relates to #662