flakes: Only ignore flakes, not all qa
Those things are somewhat justified, so we need to tell pyflakes to not
complain about them.

However, using noqa silences all linters, not just pyflakes.

Since we now use pytest-flakes when running the tests, we can benefit
from pytest's pragma feature and ask to only skip pyflakes on these
lines, rather than all linters.

This does mean those lines won't be ignored any more when running the
pyflakes command directly.

I think this is a fine trade-off for not masking other potential issues.

After all, asking a linter to ignore a line doesn't make the line good.
It still is a code smell.

One could even see this as a feature: we (and the CI) can run the tests
with pytest-flakes to keep a higher code quality, and we can still find
those code smells with the pyflakes command.