Implement user import from Llave del Saber
For the Colombian deployment, we need to be able to import user
information from their Llave del Saber system.

That requires a little data mapping, which is partially done in this
commit. However, talking this over with our Colombian friends, it seems
like the export facility in Llave del Saber is not entirely ready.

As a result, for now we'll import only the user serials, which allows at
least having the users in Ideascube. In the future, once we implement
the rest of the data mapping, running the import again will update these
users and give them all their data from Llave del Saber.

Relates to #662