Implement user import from Llave del Saber
For the Colombian deployment, we need to be able to import user
information from their Llave del Saber system.

We had started doing it, but until now we were only importing the
serials.

This commit implements the rest of the mapping so that all the user data
can be imported.

Some values don't map to values we know about, for example their
'postgrado' school level, or their 'oficios del hogar' occupation, but
that's ok, they'll be used directly.

Fixes #662