Use the `=` sign for the progress bars
That unicode black square makes for very pretty progress bars.

Until you change your font, and the square suddenly has space on its
left and right sides, and these spaces are unaccounted for when
calculating the width, and the line gets broken and it looks terrible.

In addition, we regularly have encoding errors because of it when
running the commands through an SSH connexion, with different locales
installed on the client and the server.

So let's go back to simplicity, with the default progress character.