users: Allow custom current_occupation values
Some deployments have very specific values to set here, too specific for
us to add them all as choices in our model.

This commit removes the choices constraint on the model so that any
value can be stored.

However, it also gets the form to use the new ComboBoxEntry widget, so
that a list of possible values is displayed anyway.

Custom values get stored as well, so that they appear in the list of
possible values after having been added.

Relates to #662