Allow DocumentFactory to handle preview field.
The factory.django.FileField.DEFAULT_FILENAME == 'example.dat'.
It means that by default a FileField created by factoryboy is considered as a
True value.
Before this commit, we were not defining a Document.preview field in the
factory so factoryboy created a empty FileField.
To not break the API for other tests, we need to create a "False" FileField by
default.
To do so, we need to change the DEFAULT_FILENAME to None.