Make the specimen.serial unique.
The serial number should be unique. We make this reflected in the
database schema.

The checks after the migration in the tests are not really needed.
If the `ensure_serial_is_unique` fails somehow, the `AlterField` will
fail with a integrity error as the serial is not unique.
But the checks test more than the unicity but also that just the "~number"
is added when necessary.