Add a command to remove orphan files from the mediacenter.
Orphan files are files not associated with a document in the mediacenter.
This may happen when a document is deleted or modified. In this case, the
associated file on the filesystem is not removed. This command found those
files and remove them.

We use subcommand here as the clean command will gain new features (as
remove media from the mediacenter) and it is easier to use subcommand from
the beginning instead of changing it latter.