ci: Start playing with Gitlab CI
With this, we run the tests on our deployment target: Python 3.4 on
Debian Jessie.

We use our own Docker image for this:

    https://hub.docker.com/r/ideascube/ideascube-ci

It's a simple Debian Jessie (from the official image), with the
development headers and build toolchain pre-installed.

We run the tests both in a UTF-8 and in the C locale, just like we were
doing in Travis.

Each test runs the flakes linting, which is a bit wasteful. This will
change in a subsequent commit.

Also, coverage is missing for now.

Fixes #693