Fix the serial migration
As it turns out, using null=True, blank=True and unique=True on a
Charfield is a pretty terrible idea, as the empty string used as the
default value breaks the uniqueness constraint.

Fortunately, Django devs are coming to the rescue for 1.11:

    https://code.djangoproject.com/ticket/4136

But we can't upgrade just yet, so we need another solution.

This commit ensures that ffuture Specimen objects get stored to the DB
with their serial set to None instead of the empty string.

It also modifies the migration to change all existing specimens with an
empty string as their serials to use None instead.

And then, the migration just works.