Handle our dependencies with pip-tools
This is pretty nifty.

Keeping our dependencies up to date is as simple as running:

    make upgrade-deps

And it generates **complete** pip requirements files, so that we know
exactly what gets installed, everybody (including the Debian packages)
gets the same versions of everything, no matter when they installed.

If we need to add a new dependency, or remove an old one, we do that in
the new .in files.

And from time to time, each one of us needs to sync their virtual
environment:

    make sync-deps

More details about pip-tools:

    https://github.com/nvie/pip-tools