ci: Verify code coverage
This is the last thing we were missing in the move from our Travis-based
CI to the new Gitlab one.

We have a pretty high code coverage in Ideascube, this should help
keeping it that way and improving it.

Gitlab has native integration for coverage report, so we can stop
relying on Coveralls (a proprietary platform) and still have the cute
badge.

However, we lose two things:

*   a coverage report right in the pull request page, we need to go see
    the builds associated to the pull request, one of them will have the
    coverage information.
    https://gitlab.com/gitlab-org/gitlab-ce/issues/3694

*   whether a branch increased or decreased the code coverage
    https://gitlab.com/gitlab-org/gitlab-ce/issues/17265

Both are features which might eventually come into Gitlab.

The former is a minor annoyance, but the latter is more problematic. To
try and avoid decreasing our code coverage, this commit asks pytest to
fail the build if it gets under 92%.

I hope with time we will increase the global code coverage, and with
that increase this minimal value.