ci: Ensure we didn't forget any model migration
In the past we've sometimes merged pull requests, only to realize later
that a small migration had been missed.

This mostly happens for trivial changes to the model (e.g adding a
verbose_name or a new value to a choices) or for models based on
configuration (e.g the list of languages).

With this new job in the CI, we can make sure it never happens again.

Fixes #669