ci: Add a Docker file
We've been running our continuous integration system on Travis, which
provides old Ubuntu VMs.

We'd like to start testing our software on our deployment target: Debian
Jessie.

This commit adds a Docker file to build such an image, which can then be
the basis for our CI tests.

It creates a user to run the tests as, otherwise they will be run as
root, which can have some subtle impact on some tests, leading to hard
to debug failures.

Built image will be uploaded to:

    https://hub.docker.com/r/ideascube/ideascube-ci/

Relates to #693