Bump our resumable-urlretrieve dependency
Upstream just released a new version. What's good in it is that it now
explicitly specify that it requires the typing module on Python < 3.5
(typing is part of the standard library in Python >= 3.5)

As a result, we can drop our dependency on mypy-lang. We never really
needed it, we had only added it when bringing in resumable-urlretrieve,
but that was a mistake: what we really needed was the typing module,
since resumable-urlretrieve didn't specify its dependency on it, and
mypy-lang just happened to depend on typing.