Add a command to clean tags.
Before ideascube version 0.9, we were using case-sensitive taggit.
Then at this time, taggit were creating two different tags for 'Foo' and
'foo'.

But now, with more recent versions, we are in case-insensitive. Then
taggit assumes there is only one tag for a (insensitive) string.
So if we try to tag a object with tag 'foo', taggit crashes because the
assumption is not fulfill.

This new command "tags sanitize" fixes this by merging (insensitive) equal
tags. After running it, we end with the same situation as if we were all
the time in case-insensitive setting.

This command also clean the tag names :
 - It removes punctuation characters at start and end of tag names
 - It removes tags with empty names.
 - It lowerize all tag's names.