Remove -e in requirements(-dev).txt
pip-compile doesn't support URLs as packages unless they are editable.
So we need the '-e' options in *.in files.

But, then when dh-virtualenv	install the package as editable with a
egg-link pointing to the source. As the sources change location when we
package everything, the egg-link points to nothing and we cannot import
the packages.

This is a bit silly, but the sed after the requirements.txt generation
removes the '-e' option and everything work again.
This is definitively a workaround and we should remove it once this bug is
fixed : https://github.com/spotify/dh-virtualenv/issues/200