i18n: Update the translations
This is the result of running:

    $ make collect_translations
    $ make compile_translations

Our .po/.mo files were really out of date with the source code. This
does not take the latest translations from Transifex though, which will
come in a future commit.

However, only the en locale files (the translation source) are
committed, the other locales must come from Transifex.