Keep the -e flag in requirements-dev.txt
pip-sync doesn't like URLs as requirmeents unless they use the -e flag.
We only use pip-sync on our development workstations, where we use the
requirements-dev.txt file.

In addition, the issue with dh-virtualenv which prompted us to remove
the -e option only happens when building a Debian package, which uses
the requirements.txt file.

All this means we should keep the -e in the requirements-dev.txt file,
but continue removing it in the requirements.txt file, at least until
the dh-virtualenv issue is fixed.