Properly dump the JSONField data
When using the dumpdata command, the JSONField values were just
transformed into strings.

This means that a value like:

    ['a', 'list']

would get serialized as:

    "['a', 'list']"

And then we wouldn't be able to load that, because it's not valid JSON,
because it uses single quotes to delimit strings.

This commit ensures the serialization is properly done.