Add some test data
Eventually we will be able to use this data to verify that migrations
work from the last release.

The goal is to have a CI job which runs at every pull request and:

* checks out the tag of the last release
* loads this data into the database
* checks out the branch being tested
* run the migrations

It should give us more confidence into our migrations, especially with
corner cases. This data will need to live with our releases, as
documented.

Of course, we don't have this data in a release yet, so we can't enable
a CI job. But we can do that starting with the next release.

We can also run the checks manually with a simple make target, just
like the CI would.

Relates to #670