Improve user feedback
This moves to using the typical way for Django commands to print
messages, instead of the print function. This is not only more
idiomatic, it also offers interesting possibilities for testing.

This commit also improves the English phrasing a bit, and inverts a
usage of delete/remove in a sentence to make it stick better to the
terminology used around.