Better calculate the number of remaining media
As things stand, both operations return the same result.

However, that won't be true any more once we start filtering things.

This commit just prepares us for that.

And since we're really interested in the medias installed by packages,
let's rename the variable so it expresses this, not all the left media.