Move from the setting files for IDEASCUBE_NAME
We've had the server.site-name in-database configuration for a few
releases now, and ansiblecube sets it for new boxes.

This means the time has come to start removing it from the settings
files, to eventually remove those settings files entirely.

But before removing them from the settings files, the first thing to get
old boxes to have the configuration set in their database.

This commit adds a migration which searches for an IDEASCUBE_NAME in the
settings. If it finds one, and there was no in-database server-name set,
then the configuration gets set to what was in the settings file.

Once a box will have been upgraded to a release including this
migration, then we can remove the IDEASCUBE_NAME option from their
settings file starting with the next release.