conf: Split the HOME_CARDS
We really have two types of home cards: the ones corresponding to apps
built into Ideascube, and the ones corresponding to external apps which
must be installed on the box.

This commit splits them into two distinct settings, which makes it
easier for box-specific settings files to override one or the other.