Add translation for tinyMCE.
There is only tree translation made for tinyMCE we also use `ar`, `es` and
`fr_FR`.

Other languages we use (`am`, `bm`, `so`, `sw`) are not available for
tinyMCE. So, we still use the english version. If we want them, we will
have to translate it in the tinyMCE project.

As we use the `fr` code to represent the french language in ideascube,
we need a map to somehow pass `fr_FR` to tinyMCE.