Downgrade our python-NetworkManager dependency
The latest releases are giving us trouble. Version 2.0.0 is a full
rewrite, which seems to have broken compatibility with the old
NetworkManager on Debian Jessie.

Git master has a patch to restore compatibility:

    https://github.com/seveas/python-networkmanager/commit/cad4dc79efd8d29e9fe1f1cd03f76e61534af808

However, even with that applied we still seem to encounter issues. So
for now, we've decided to stick with the older version.