Instantiate the catalog only once for the home page view.
Instantiation of the `Catalog` can be time consuming. (Until 4 secondes
on koombook machine).

By keeping a instance of the catalog in the global namespace, we avoid
us to create a new catalog instance at each request.

The drawback is that any changes made in the catalog (installation or
removal of packages) in not take into account by ideascube.
We have to restart ideascube to reflect the new cards to display.