Better document creation in test.
We don't need to store the test media files in sub-directory. If we give
a path with sub-directory as the `File` name, django will store the file
in the sub-directory, wherever the original were stored.
However, we need to keep the data files because we use them in the
`import_media` tests.

We also move the document creation in a dedicated function to avoid code
duplication.