ci: Use the commit hash, not the branch
Turns out, Gitlab CI doesn't checkout the branch for a merge request, it
checks out the commit (and thus is in a detached HEAD).

As a result, when we check out the branch further down, we can end up on
an older version of the branch (e.g if the branch had been force-pushed)
because Gitlab CI doesn't really start from a clean container every
time, but keeps some cache.