Move to Django 1.11
One of the major new things introduced in Django 1.11 is template-based
rendering for widgets.

As a result, our custom LangSelect widget needs to be adapted to the
change.

The old render_option method now becomes simply a local option template.

The old render_options method however is a bit more complex, and we need
to move its filtering to the new optgroups method.