ci: Massively speed up the job runs
We currently install the dependencies in the virtual environment when
running the CI jobs. That is, each job will download and install the
dependencies with pip at the beginning.

But those dependencies are the same for all the tasks.

This commit speeds everything up by pre-installing the dependencies in
the Docker images.

The tasks wil then just synchronize their dependencies when they start,
which in most cases will be almost instantaneous, as long as we keep our
Docker images up to date.