ci: Properly run the C locale tests
Click is one of our dependencies, through pip-tools. Click does not like
being used on a non-UTF8 locale.

Now that we synchronize the dependencies in CI jobs with pip-tools
instead of installing them with pip, Click gets used in the C locale and
errors out, failing the CI jobs.

This commit changes the job definition so that only the actual tests are
run with the C locale, not the whole task.