Move to django-multiselectfield
The django-select-multiple-field module we were using seems completely
abandoned upstream, as evidenced by the absence of reaction on the
discussion about Django 1.10+ compatibility:

    https://github.com/kelvinwong-ca/django-select-multiple-field/pull/18

We've been using a fork for some time now, and overall it works fine,
but relying on a local fork of an unmaintained project is a bad idea, as
we'd have to assume the maintainance.

Fortunately, there exists a maintained module which does very similar
things: django-multiselectfield

And in fact, what it does differently is what we wanted! So we can drop
all our custom code in the CommaSeparatedCharField class and go
completely with this new module.