Add a "suitable for all languages" option
Some of the content users put in Ideascube is not related to a specific
language, for example pictures, music-only media, etc...

This new option allows expressing this fact.

The question thus becomes what code to use to represent this.

The obvious answers were "all" or a "*" wildcard. We decided against the
former because it might in theory become an actual language code in the
ISO standard, and our admin was vehemently against the latter because it
would be harder to escape when grepping.

ISO language codes can only be letters, so if we use a number we can be
sure we will never conflict with anything.

L33t-speak to the rescue! We can use "411" as a reference to "All".

Relates to #377