Capitalize the language filters
The language local names respect the grammar rules for capitalizing
language names in each language, for example:

    I speak English.

But:

    Je parle français.

This is obviously a good thing, when using those language names inside
sentences.

However, here we're using them on their own, as button labels, so we
should capitalize them all, both for consistency and because it is the
correct thing to do.