Stop explicitly depending on html5lib from git
We had to do that so we would get a newer html5lib which had dropped its
dependency on setuptools, as that cause issues with our usage of
pip-tools.

But since pip-tools 1.9.0, we don't need to do that any more:
pip-compile now properly excludes setuptools and its dependencies.

We don't use html5lib directly, it is a transitive dependency through
bleach, so we can now completely remove it from our requirements.in file
and let pip-compile do its magic.