Move `heading` block in a specific row.
The form is composed of two different columns.
If we keep the heading in the main column, the two columns are not aligned
(The side column is aligned with the heading).

By moving the heading outside of the column in a specific row, the two
"content" columns are now aligned.