Move form's buttons in the main content column.
The main content column is used for main fields and side column to
"metadata" fields.

If we keep the submit buttons in a specific row and if the side column has a lot of fields, the submit button may be pushed
down the page.

By moving the buttons in the main column, they always are just after the
main field.