Use a margin instead of a padding to separate two columns.
This is needed for elements with a background (as button) or a border.
The `padding` in inside the bounding-box and the `margin` is outside.
With a `padding` the background is "extended" on the left.

There is potential draw back because the different padding are added
where the margin are merged (to keep only the biggest).
This is not important here as we are specifying only the left `margin`
and only between two `col` elements, so there will not be two `margin` to
merge.