Split the form's fields in the two columns.
Use the main column for main fields and side column for "metadata" fields.

It's reduce the form size and make the accent on the main content instead
of "hiding" it in between less important fields.

To do so, this commit creates a new template `form-sidebar.html` who defines
two blocks to be completed by inherited template :

 - `form` for the main fields (by default, all fields of the form)
 - `sidebar` for "medatada" fields (empty by default)