Move side column on the left.
We use to see the main content on the right and have secondary information
on the left marge.
And I suppose it is the same thing (but opposite) for RTL users.

So declare the side column first to display it on left side
(right side for RTL).

Now that side column is declared first, we also need to hide it in form
where we don't use the side column.