home: Simplify the settings some more
We have several setting options to control the cards visible on the home
page.

One of them, EXTRA_APP_CARDS control the visibility of some extra
applications which can be installed along with Ideascube.

It basically only lists card identifiers, but does so in a format which
is more complex than it needs to be, for historical reasons.

This commit simplifies the setting files, moving the complexity to the
code.