home: Simplify the settings
We have several setting options to control the cards visible on the home
page.

One of them, BUILTIN_APP_CARDS control the visibility of the
applications built into Ideascube.

It basically only lists card identifiers, but does so in a format which
is more complex than it needs to be, for historical reasons.

This commit simplifies the setting files, moving the complexity to the
code.