home: Factorize some templates
This commit introduces a new class to represent some home cards, more
specifically the ones related to builtin apps.

Thanks to this, we can completely drop the 3 different templates and
just use a common one.

Of course, this is only a first step, future commits will move more
types of cards to this new representation.