We do not have any more dependency requiring git
This makes for smaller Docker images, and faster sbuild builds, as Git
brings along quite a lot of Perl packages.

We might end up requiring it in the future for some new dependencies,
but we should try as hard as possible not to rely on unreleased
software, and we can always add it back if needed.