debian: Ensure the build happens in a UTF-8 locale
Python uses a default encoding which is based on the current locale.

Before Python 3.7, if that locale is POSIX, which is what happens when
building Debian packages in a chroot, then the encoding is 'ascii'.

This obviously breaks when trying to `pip install` some dependency and
for example the author name in the setup.py uses non-ascii characters.

We could move to Python 3.7 right now, which fallsback to using C.UTF-8
instead of POSIX so that the encoding will be utf-8.

Hahahaha, of course we can't. :)

This commit tells the Debian package tooling to use the C.UTF-8 locale
during the build, whoch fixes the problem.