home: Improve handling of the custom cards
This allows reusing the same 'external' template as used elsewhere. At
this point we only have 3 card templates left!

In addition to having less templates to maintain, this allows setting
HTML descriptions for custom cards, which wasn't possible before.

I'm also reordering the comments so that they correspond to the order of
the attributes in the ideascube.cards.Card class, because I'm a pedant.