debian: Require libyaml-dev when building
We use PyYAML quite extensively in Ideascube, especially in the catalog.
But the catalog is used to generate the home page, so that's quite
critical.

PyYAML has an optional C extension to speed things up, but it's only
enabled at build-time if the libyaml compilation header are available.

This commit adds the build requirement for the Debian package, so that
the PyYAML C extension gets built in our package, which should speed up
the catalog.