Remove the touch-reload line
This does not actually work.

The file is a symlink to a configuration file, which gets handled
differently by dpkg.

As a result the mtime does not change when upgrading the package, and so
uwsgi does not reload the app.

Since it doesn't work, and since we already restart uwsgi in ansiblecube
anyway, let's remove this.