catalog: Improve some output messages
This moves some lines of code for seemingly no reason, but it helps
reading this whole thing if printing and adding the used handlers are
kept close to the actual action in the try block.

Some messages are also made more concise and with more information.

In addition, this better uses string formatting to prepare for
f-strings if we ever got to move to Python 3.6.