catalog: Upgrading means installing the latest version
When an admin asks to upgrade the package foo, all they care about is
that after the command has run, the latest version of foo is installed.

As a result, we shouldn't fail if a package isn't installed when an
admin requests it be updated, we should just install the latest version.

Fixes #802