Move to the new django API for the i18n javascript catalog.
The `JavaScriptCatalog` will try to load translation in `LOCALE_PATHS`
directories and in `locale` directories of specified packages.

But if `packages` argument is not specified or `bool(packages) == False`
all `INSTALLED_APPS` will be considered.

So, we need to specify that we load translation for `ideascube`
application only. And we don't need to specify `LOCALE_PATH` anymore.