Load jsi18n only when necessary.
The jsi18n is only used in javascript and the only javascript we have
is `ideascube.js` used in forms and `document_select.html`.

So load `jsi18n` only there. We do not need to load it all the time.

This should greatly reduce the impact of #815 for most of our pages.