Bring back LOCALE_PATHS
We removed it recently, in commit 9af4650, assuming it was not
necessary.

And indeed, it is not necessary for the locales known by Django.

Unfortunately, we add some locales which Django doesn't know of, and
these were broken by this change: it became impossible to switch to
these locales, as Django would not find the corresponding language codes
in its own internal folder.

This commit makes them work again, by telling Django to search for those
language codes in our locale folder, in addition to its own.

Relates to #819