Switch to the fa-ir language code
Django doesn't use locales for the translations, it uses language codes
of the form xx or xx-yy, where xx is the language code and yy is the
region code.

This differs from locale codes which are of the form xx_YY.

This commit makes us use the fa-ir language code instead of the fa_IR
locale in the configuration.

However, when it comes to actually loading the translations (.mo file)
on the disk, Django converts the language code to a locale one. As a
result, ideascube/locale/fa_IR/ stays unchanged.

Fixes #819