Get the tests to display warnings
Since its 1.11 release, Django doesn't report deprecation warnings any
more by default. This means we missed some after the upgrade.

It is especially bad because Django 2.0 will break a lot of things, and
Django 1.11 is meant as a stepping stone towards that, adding lots of
deprecationw warnings to help migrate to 2.0.

With pytest 3.1, we can control finely what warnings we want to display,
or even whether they should be considered failures. Upgrading makes it
trivial to display the deprecation warnings.

In this way, the CI will help us move towards Django 2.0, or even
upgrade any of our dependencies which handle compatibility breaks with
deprecation periods, the way Django does.

Relates to #817