Explicitly set on_delete for old migrations
Django 2.0 will make it mandatory, so we need to do it now, using
CASCADE to retain the current default behaviour.

The blog one is tricky, because as of today the blog model uses PROTECT,
not CASCADE. However, we're only modifying the **initial** migration
which created the model here, and at the time the model was using
CASCADE, and moved to PROTECT in a subsequent migration is already
correct.

https://docs.djangoproject.com/en/1.11/ref/models/fields/#django.db.models.ForeignKey.on_delete