Use the thumbnail coming from the catalog if present.
The catalog can now declare a thumbnail to use for a package.

If this thumbnail is present, we store it in the mediacenter as a hidden
document and we use it in the home cards.

We use a "complex" structure to store the thumbnail in the catalog
(the `thumbnail` entry is a dict containing the `data` entry) to allow
potential future extensions (`url` or `mime-type` entries or different
sizes)