Some misc clean up in the css.
The `background-position` is wrong here. Those two settings are here
almost from the beginning and we've rewrite the card template few time
since. By forcing the `top` position, we prevent the image to be centered.

The `#cards td label` is very old also and was used to make the label
in the card `clickable`. Now that the whole card is a link, this is
no more necessary (and already unused).