Catalog can now "standalone" zim packages.
Now that kiwix start to generate zim file with embeded fulltext index
(and no more generate "portable" archives), we can start (and need) to
install "packages" that are just a zim file.

Nothing complicated to do, just copy the zim file at the right place.

The `type` of the package to use in the catalog is `zim`.