Stop updating the catalog cache when managing remotes
Back in the day, we thought the main interaction would be to add a
remote on a machine, and then just try installing contents from it. In
that paradigm, it made sense to automatically update the cache when
adding or removing a remote.

However, what actually happens is that Ansiblecube adds all the remotes
one after the other, which means the cache gets updated after each
addition.

This essentially reverts 577941b57a054fd70136c8dc869ab84dd9d674e0, as we
found this was actually detrimental.

Fixes #824