Stop keeping the packages after install/upgrade
We implement things this way, keeping the package in the cache after
installing it, because the idea was that Koombooks would share their
caches. That is, you'd install some packages on a Koombook, which could
then be used as an external cache (--package-cache) for another one.

This never actually got used this way, and instead we end up filling the
disks of the Koombooks with the package caches.

The default behaviour should reflect the common use-case, and as such
with this commit we remove the downloaded packages after they were
successfully installed, upgraded or reinstalled.

We also add an option to keep them, in case admins explicitly want to.

Fixes #820